requirements.txt
# Python dependencies for phishing detection benchmarking
# Install with: uv pip install -r requirements.txt

# Core ML/DL
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Energy tracking
codecarbon>=2.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0

# Optional: for Jupyter notebook exploration
# jupyter>=1.0.0
# ipywidgets>=8.0.0
